fixed split window view for MEI editing in recognition mode; changed FacsTable entries to include all 4 coordinates and made MusStaff::ecart a double instead of int for more accurate vertical placement of staves drawn from MEI

+++ src/mus/musiobin.cpp
@@ -174,7 +174,7 @@ bool MusBinOutput::WriteStaff( const MusStaff *staff )
 	Write( &staff->notAnc, 1 );
 	Write( &staff->grise, 1 );
 	Write( &staff->invisible, 1 );
-	uint16 = wxUINT16_SWAP_ON_BE( staff->ecart );
+	uint16 = wxUINT16_SWAP_ON_BE( (unsigned short)staff->ecart );
 	Write( &uint16, 2 );
 	Write( &staff->vertBarre, 1 );
 	Write( &staff->brace, 1 );
